+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+## 2.17.0
+- Added custom MoveDevice display name.
+- Added MoveDeviceState. Ensure synchronized option.
+- Live Location TAG feature setLiveLocationTag(...) deprecated
+
 ## 2.16.0
 - Added getMoveConfig.
-- Added iBeacon service registration.
 - Stability improvements.
 - Bug fixes.
 - Split iOS HealthKit dependent APIs into separate packages. See documentation if required.